add test for string type

+++ lib/global/convert.cc
@@ -2,7 +2,7 @@
 
 std::string vectorUInt8ToString(std::vector<uint8_t> vectorIn) {
   std::string stringResult = "";
-  for (uint16_t i = 0; i < vectorIn.size(); i++) {
+  for (uint32_t i = 0; i < vectorIn.size(); i++) {
     stringResult += vectorIn[i];
   }
   return stringResult;
@@ -10,7 +10,7 @@ std::string vectorUInt8ToString(std::vector<uint8_t> vectorIn) {
 
 std::vector<uint8_t> stringToVectorUInt8(std::string stringIn) {
   std::vector<uint8_t> vectorResult;
-  for (uint16_t i = 0; i < stringIn.length(); i++) {
+  for (uint32_t i = 0; i < stringIn.length(); i++) {
     vectorResult.push_back(stringIn[i]);
   }
   return vectorResult;
